Sharing test cases for DefaultApiClientSpec and BlockingApiClientSpec
